refactor: Answer 삭제 시 요청자 검증 로직 이동(Question -> Answers)

+++ src/main/java/nextstep/qna/domain/Answers.java
@@ -1,5 +1,6 @@
 package nextstep.qna.domain;
 
+import nextstep.qna.CannotDeleteException;
 import nextstep.users.domain.NsUser;
 
 import java.util.ArrayList;
@@ -34,7 +35,10 @@ public class Answers {
         return false;
     }
 
-    public List<DeleteHistory> delete() {
+    public List<DeleteHistory> delete(NsUser loginUser) throws CannotDeleteException {
+        if (!isAnswerOwner(loginUser)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
         List<DeleteHistory> deleteHistories = new ArrayList<>();
         for (Answer answer : values) {
             deleteHistories.add(answer.delete());
